+++ apps/sasquatch/src/main/java/com/microsoft/appcenter/crashes/CrashesPrivateHelper.java
@@ -5,18 +5,9 @@
 
 package com.microsoft.appcenter.crashes;
 
-import com.microsoft.appcenter.crashes.ingestion.models.ErrorAttachmentLog;
-
-import java.util.Map;
-
 public final class CrashesPrivateHelper {
 
     private CrashesPrivateHelper() {
-    }
-
-    public static void trackException(Throwable throwable, Map<String, String> properties, Iterable<ErrorAttachmentLog> attachments) {
-        // TODO add attachments param
-        Crashes.trackException(throwable, properties);
     }
 
     public static void saveUncaughtException(Thread thread, Throwable exception) {