+++ sdk/appcenter-crashes/src/main/java/com/microsoft/appcenter/crashes/Crashes.java
@@ -246,12 +246,11 @@ public class Crashes extends AbstractAppCenterService {
 
     /**
      * Track an exception.
-     * TODO the backend does not support that service yet, will be public method later.
      *
      * @param throwable An exception.
      */
     @SuppressWarnings("SameParameterValue")
-    static void trackException(@NonNull Throwable throwable) {
+    public static void trackException(@NonNull Throwable throwable) {
         trackException(throwable, null);
     }
 
@@ -262,12 +261,11 @@ public class Crashes extends AbstractAppCenterService {
      * The properties keys can not be null or empty, maximum allowed key length = 64.
      * The properties values can not be null, maximum allowed value length = 64.
      * Any length of name/keys/values that are longer than each limit will be truncated.
-     * TODO the backend does not support that service yet, will be public method later.
      *
      * @param throwable  An exception.
      * @param properties Optional properties.
      */
-    static void trackException(@NonNull Throwable throwable, Map<String, String> properties) {
+    public static void trackException(@NonNull Throwable throwable, Map<String, String> properties) {
         Map<String, String> validatedProperties = ErrorLogHelper.validateProperties(properties, "HandledError");
         getInstance().queueException(throwable, validatedProperties);
     }